# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TempNest API - Temporary Email Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileCurrentPlan = typeof UserProfileCurrentPlan[keyof typeof UserProfileCurrentPlan];


export const UserProfileCurrentPlan = {
  free: 'free',
  pro: 'pro',
  business: 'business',
} as const;

/**
 * @nullable
 */
export type UserProfileNextPlan = typeof UserProfileNextPlan[keyof typeof UserProfileNextPlan] | null;


export const UserProfileNextPlan = {
  free: 'free',
  pro: 'pro',
  business: 'business',
} as const;

export type UserProfileStatus = typeof UserProfileStatus[keyof typeof UserProfileStatus];


export const UserProfileStatus = {
  active: 'active',
  suspended: 'suspended',
  banned: 'banned',
} as const;

export interface UserProfile {
  id: number;
  firebaseUid: string;
  email: string;
  name: string;
  currentPlan: UserProfileCurrentPlan;
  /** @nullable */
  nextPlan?: UserProfileNextPlan;
  /** @nullable */
  planStartDate?: string | null;
  /** @nullable */
  planExpiryDate?: string | null;
  planRenewalReminder?: boolean;
  credits: number;
  maxCredits?: number;
  dailyRefill?: number;
  activeInboxCount?: number;
  maxInboxes?: number;
  status?: UserProfileStatus;
  isAdmin?: boolean;
  notifyNewEmail?: boolean;
  notifyOtp?: boolean;
  notifyLowCredits?: boolean;
  notifyWeeklySummary?: boolean;
  createdAt: string;
}

export interface UserProfileUpdate {
  name?: string;
}

export interface NotificationPreferences {
  notifyNewEmail?: boolean;
  notifyOtp?: boolean;
  notifyLowCredits?: boolean;
  notifyWeeklySummary?: boolean;
}

export interface Email {
  id: string;
  inboxId: string;
  from: string;
  subject: string;
  preview?: string;
  isRead: boolean;
  hasOtp?: boolean;
  /** @nullable */
  otpCode?: string | null;
  createdAt: string;
}

export interface DashboardSummary {
  credits: number;
  maxCredits?: number;
  inboxCount: number;
  maxInboxes?: number;
  emailsToday: number;
  otpsExtracted: number;
  creditsUsedToday: number;
  recentEmails?: Email[];
}

export interface Inbox {
  id: string;
  address: string;
  domain: string;
  isActive: boolean;
  isPriority?: boolean;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
  emailCount: number;
  unreadCount: number;
}

export interface InboxInput {
  /** @nullable */
  customName?: string | null;
  isPriority?: boolean;
}

export interface InboxRefreshResult {
  newEmails: number;
  inbox: Inbox;
}

export interface EmailDetail {
  id: string;
  inboxId: string;
  from: string;
  subject: string;
  /** @nullable */
  bodyText?: string | null;
  /** @nullable */
  bodyHtml?: string | null;
  isRead: boolean;
  hasOtp?: boolean;
  /** @nullable */
  otpCode?: string | null;
  createdAt: string;
}

export interface MarkReadInput {
  isRead: boolean;
}

export interface OtpExtraction {
  found: boolean;
  /** @nullable */
  code?: string | null;
  /** @nullable */
  length?: number | null;
}

export interface OtpRecord {
  id: number;
  emailId: string;
  code: string;
  length?: number;
  extractedAt: string;
}

export interface CreditWallet {
  balance: number;
  maxBalance: number;
  dailyRefill: number;
  /** @nullable */
  lastRefillAt: string | null;
}

export type CreditTransactionType = typeof CreditTransactionType[keyof typeof CreditTransactionType];


export const CreditTransactionType = {
  debit: 'debit',
  credit: 'credit',
  refund: 'refund',
  purchase: 'purchase',
  refill: 'refill',
} as const;

export interface CreditTransaction {
  id: number;
  type: CreditTransactionType;
  amount: number;
  description: string;
  balanceAfter?: number;
  createdAt: string;
}

export interface Plan {
  id: string;
  name: string;
  price: number;
  currency?: string;
  billingPeriod?: string;
  credits: number;
  maxInboxes: number;
  maxCredits?: number;
  dailyRefill?: number;
  features: string[];
  /** @nullable */
  stripePriceId?: string | null;
}

export type SubscriptionStatus = typeof SubscriptionStatus[keyof typeof SubscriptionStatus];


export const SubscriptionStatus = {
  active: 'active',
  cancelled: 'cancelled',
  past_due: 'past_due',
  free: 'free',
} as const;

export interface Subscription {
  planId: string;
  planName: string;
  /** @nullable */
  nextPlanId?: string | null;
  /** @nullable */
  nextPlanName?: string | null;
  status: SubscriptionStatus;
  /** @nullable */
  currentPeriodEnd: string | null;
  /** @nullable */
  planStartDate: string | null;
  /** @nullable */
  planExpiryDate: string | null;
  cancelAtPeriodEnd?: boolean;
}

export type ScheduleDowngradeInputPlanId = typeof ScheduleDowngradeInputPlanId[keyof typeof ScheduleDowngradeInputPlanId];


export const ScheduleDowngradeInputPlanId = {
  free: 'free',
  pro: 'pro',
  business: 'business',
} as const;

export interface ScheduleDowngradeInput {
  planId: ScheduleDowngradeInputPlanId;
}

export type CheckoutInputType = typeof CheckoutInputType[keyof typeof CheckoutInputType];


export const CheckoutInputType = {
  credits: 'credits',
  subscription: 'subscription',
} as const;

export interface CheckoutInput {
  type: CheckoutInputType;
  /** @nullable */
  planId?: string | null;
  /** @nullable */
  creditPackId?: string | null;
}

export interface CheckoutSession {
  url: string;
}

export interface DailyUsage {
  date: string;
  emailsReceived: number;
  inboxesCreated: number;
  creditsUsed: number;
  otpsExtracted: number;
}

export interface UsageTotals {
  totalEmails: number;
  totalInboxes: number;
  totalCreditsUsed: number;
  totalOtps: number;
}

export interface UsageAnalytics {
  daily: DailyUsage[];
  totals: UsageTotals;
}

export interface AdminStats {
  totalUsers: number;
  activeUsers: number;
  totalInboxes: number;
  totalEmails: number;
  totalRevenue: number;
  creditsIssued: number;
  proUsers?: number;
  businessUsers?: number;
  newUsersToday?: number;
  emailsToday?: number;
}

export interface AdminUser {
  id: number;
  firebaseUid: string;
  email: string;
  name: string;
  plan: string;
  credits: number;
  status: string;
  inboxCount?: number;
  emailCount?: number;
  isAdmin?: boolean;
  createdAt: string;
}

export interface AdminUserPage {
  users: AdminUser[];
  total: number;
  page: number;
  pageSize: number;
}

export type AdminUserUpdateStatus = typeof AdminUserUpdateStatus[keyof typeof AdminUserUpdateStatus];


export const AdminUserUpdateStatus = {
  active: 'active',
  suspended: 'suspended',
  banned: 'banned',
} as const;

export type AdminUserUpdatePlan = typeof AdminUserUpdatePlan[keyof typeof AdminUserUpdatePlan];


export const AdminUserUpdatePlan = {
  free: 'free',
  pro: 'pro',
  business: 'business',
} as const;

export interface AdminUserUpdate {
  status?: AdminUserUpdateStatus;
  grantCredits?: number;
  plan?: AdminUserUpdatePlan;
}

export interface AdminInboxItem {
  id: string;
  address: string;
  userId: number;
  userEmail?: string;
  isActive: boolean;
  emailCount: number;
  createdAt: string;
}

export interface AdminInboxPage {
  inboxes: AdminInboxItem[];
  total: number;
  page: number;
  pageSize: number;
}

export interface MonthlyRevenue {
  month: string;
  revenue: number;
  transactions: number;
}

export interface PlanDistribution {
  plan: string;
  count: number;
  percentage: number;
}

export interface AdminRevenue {
  totalRevenue: number;
  monthly: MonthlyRevenue[];
  planDistribution?: PlanDistribution[];
}

export type ScheduleDowngrade200 = {
  success: boolean;
  currentPlanId: string;
  nextPlanId: string;
  /** @nullable */
  planExpiryDate: string | null;
};

export type ListAdminUsersParams = {
search?: string;
page?: number;
};

export type ListAdminInboxesParams = {
page?: number;
};