# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TempNest API - Temporary Email Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inboxes:
    get:
      operationId: listInboxes
      tags: [inboxes]
      summary: List user's inboxes
      responses:
        "200":
          description: List of inboxes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inbox"
        "401":
          description: Unauthorized
    post:
      operationId: createInbox
      tags: [inboxes]
      summary: Create a new temporary inbox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InboxInput"
      responses:
        "201":
          description: Created inbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inbox"
        "400":
          description: Insufficient credits or daily limit reached
        "401":
          description: Unauthorized
...
  /emails/{emailId}:
    get:
      operationId: getEmail
      tags: [emails]
      summary: Get email details
      parameters:
        - name: emailId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Email details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailDetail"
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /emails/{emailId}/read:
    patch:
      operationId: markEmailRead
      tags: [emails]
      summary: Mark email as read
      parameters:
        - name: emailId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkReadInput"
      responses:
        "200":
          description: Updated email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
        "401":
          description: Unauthorized
  /emails/{emailId}/otp:
    get:
      operationId: extractEmailOtp
      tags: [emails]
      summary: Extract OTP from email body
      parameters:
        - name: emailId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Extracted OTP result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpExtraction"
        "401":
          description: Unauthorized
        "404":
          description: Not found

  # ── Credits ──────────────────────────────────────────────────────
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard statistics
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserPage"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/users/{userId}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update a user (suspend, ban, grant credits)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Delete a user (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/inboxes:
    get:
      operationId: listAdminInboxes
      tags: [admin]
      summary: List all inboxes (admin)
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: All inboxes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInboxPage"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/revenue:
    get:
      operationId: getAdminRevenue
      tags: [admin]
      summary: Get revenue analytics (admin)
      responses:
        "200":
          description: Revenue analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRevenue"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
components:
  schemas:
...
    Inbox:
      type: object
      required: [id, address, domain, isActive, createdAt, emailCount, unreadCount]
      properties:
        id:
          type: string
        address:
          type: string
        domain:
          type: string
        isActive:
          type: boolean
        isPriority:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
        emailCount:
          type: integer
        unreadCount:
          type: integer
    InboxInput:
      type: object
      properties:
        customName:
          type: ["string", "null"]
        isPriority:
          type: boolean
...
    Email:
      type: object
      required: [id, inboxId, from, subject, isRead, createdAt]
      properties:
        id:
          type: string
        inboxId:
          type: string
        from:
          type: string
        subject:
          type: string
        preview:
          type: string
        isRead:
          type: boolean
        hasOtp:
          type: boolean
        otpCode:
          type: ["string", "null"]
        createdAt:
          type: string
    EmailDetail:
      type: object
      required: [id, inboxId, from, subject, isRead, createdAt]
      properties:
        id:
          type: string
        inboxId:
          type: string
        from:
          type: string
        subject:
          type: string
        bodyText:
          type: ["string", "null"]
        bodyHtml:
          type: ["string", "null"]
        isRead:
          type: boolean
        hasOtp:
          type: boolean
        otpCode:
          type: ["string", "null"]
        createdAt:
          type: string
    MarkReadInput:
      type: object
      required: [isRead]
      properties:
        isRead:
          type: boolean

    # ── OTP ─────────────────────────────────────────────────────────
    OtpExtraction:
      type: object
      required: [found]
      properties:
        found:
          type: boolean
        code:
          type: ["string", "null"]
        length:
          type: ["integer", "null"]
...
    AdminStats:
      type: object
      required: [totalUsers, activeUsers, totalInboxes, totalEmails, totalRevenue, creditsIssued]
      properties:
        totalUsers:
          type: integer
        activeUsers:
          type: integer
        totalInboxes:
          type: integer
        totalEmails:
          type: integer
        totalRevenue:
          type: number
        creditsIssued:
          type: integer
        proUsers:
          type: integer
        businessUsers:
          type: integer
        newUsersToday:
          type: integer
        emailsToday:
          type: integer
    AdminUser:
      type: object
      required: [id, firebaseUid, email, name, plan, credits, status, createdAt]
      properties:
        id:
          type: integer
        firebaseUid:
          type: string
        email:
          type: string
        name:
          type: string
        plan:
          type: string
        credits:
          type: integer
        status:
          type: string
        inboxCount:
          type: integer
        emailCount:
          type: integer
        isAdmin:
          type: boolean
        createdAt:
          type: string
    AdminUserPage:
      type: object
      required: [users, total, page, pageSize]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    AdminUserUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, suspended, banned]
        grantCredits:
          type: integer
        plan:
          type: string
          enum: [free, pro, business]
    AdminInboxPage:
      type: object
      required: [inboxes, total, page, pageSize]
      properties:
        inboxes:
          type: array
          items:
            $ref: "#/components/schemas/AdminInboxItem"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    AdminInboxItem:
      type: object
      required: [id, address, userId, isActive, emailCount, createdAt]
      properties:
        id:
          type: string
        address:
          type: string
        userId:
          type: integer
        userEmail:
          type: string
        isActive:
          type: boolean
        emailCount:
          type: integer
        createdAt:
          type: string
    AdminRevenue:
      type: object
      required: [totalRevenue, monthly]
      properties:
        totalRevenue:
          type: number
        monthly:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyRevenue"
        planDistribution:
          type: array
          items:
            $ref: "#/components/schemas/PlanDistribution"
    MonthlyRevenue:
      type: object
      required: [month, revenue, transactions]
      properties:
        month:
          type: string
        revenue:
          type: number
        transactions:
          type: integer
    PlanDistribution:
      type: object
      required: [plan, count, percentage]
      properties:
        plan:
          type: string
        count:
          type: integer
        percentage:
          type: number