# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TempNest API - Temporary Email Platform
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: number;
  firebaseUid: string;
  email: string;
  name: string;
  plan: string;
  credits: number;
  status: string;
  inboxCount?: number;
  emailCount?: number;
  isAdmin?: boolean;
  createdAt: string;
}